Generic functions for LOO predictions

See the methods in the rstanarm package for examples.

loo_linpred(object, ...)

loo_predict(object, ...)

loo_predictive_interval(object, ...)

loo_pit(object, ...)

# S3 method for default
loo_pit(object, y, lw, ...)

Arguments

object
The object to use.

...
Arguments passed to methods. See the methods in the rstanarm package for examples.

y
For the default method of loo_pit, a vector of \(y\) values the same length as the number of columns in the matrix used as object.

lw
For the default method of loo_pit, a matrix of log-weights of the same length as the number of columns in the matrix used as object.

Value

loo_predict, loo_linpred, and loo_pit (probability integral transform) methods should return a vector with length equal to the number of observations in the data. loo_predictive_interval methods should return a two-column matrix formatted in the same way as for predictive_interval.

See also

The rstanarm package for example methods (CRAN, GitHub).

The guidelines for developers of R packages interfacing with Stan, a copy of which can be found in the package vignettes. See browseVignettes("rstantools") or vignette(package = "rstantools"). The document is also available online at the rstantools page on the CRAN website.