Generic function for predictive intervals

See predictive_interval.stanreg in the rstanarm package for an example.

predictive_interval(object, ...)

# S3 method for default
predictive_interval(object, prob = 0.9, ...)

Arguments

object
The object to use.

...
Arguments passed to methods. See the methods in the rstanarm package for examples.

prob
A number \(p \in (0,1)\) indicating the desired probability mass to include in the intervals.

Value

predictive_interval methods should return a matrix with two columns and as many rows as data points being predicted. For a given value of prob, \(p\), the columns correspond to the lower and upper \(100p\)% interval limits and have the names \(100\alpha/2\)% and \(100(1 - \alpha/2)\)%, where \(\alpha = 1-p\). For example, if prob=0.9 is specified (a \(90\)% interval), then the column names would be "5%" and "95%", respectively.

The default method just takes object to be a matrix and computes quantiles, with prob defaulting to 0.9.

See also

The rstanarm package for example methods (CRAN, GitHub).

The guidelines for developers of R packages interfacing with Stan, a copy of which can be found in the package vignettes. See browseVignettes("rstantools") or vignette(package = "rstantools"). The document is also available online at the rstantools page on the CRAN website.

Examples

# Default method takes a numeric matrix (of draws from posterior
# predictive distribution)
ytilde <- matrix(rnorm(100 * 5, sd = 2), 100, 5) # fake draws
predictive_interval(ytilde, prob = 0.8)

#>            10%      90%
#> [1,] -2.462452 1.875353
#> [2,] -2.003644 1.919153
#> [3,] -2.383818 2.475257
#> [4,] -3.157015 2.554874
#> [5,] -2.848931 2.617529

# Also see help("predictive_interval", package = "rstanarm")