prior_summary {rstantools}	R Documentation

Generic function for extracting information about prior distributions

Description

See prior_summary.stanreg in the rstanarm package for an example.

Usage

prior_summary(object, ...)

## Default S3 method:
prior_summary(object, ...)

Arguments

object
The object to use.

...
Arguments passed to methods. See the methods in the rstanarm package for examples.

Value

prior_summary methods should return an object containing information about the prior distribution(s) used for the given model. The structure of this object will depend on the method.

The default method just returns object$prior.info, which is NULL if there is no 'prior.info' element.

See Also

The rstanarm package for example methods (CRAN, GitHub).

Guidelines and recommendations for developers of R packages interfacing with Stan and a demonstration getting a simple package working can be found in the vignettes included with rstantools and at http://mc-stan.org/rstantools/articles/.

Examples

# See help("prior_summary", package = "rstanarm")

[Package rstantools version 1.5.1 Index]