Create the skeleton of a new R package with Stan programs

The rstan_package_skeleton function helps get you started developing R packages that interface with Stan via the rstan package. rstan_package_skeleton is very similar to package.skeleton but is designed for source packages that want to include Stan Programs that can be built into binary versions (i.e., pre-compiled like rstanarm). See Details for a few ways that it differs from package.skeleton.

See the See Also section below for links to recommendations for developers and a step by step walkthrough of what to do after running rstan_package_skeleton.

rstan_package_skeleton(name = "anRpackage", list = character(),
  environment = .GlobalEnv, path = ".", force = FALSE,
  code_files = character(), stan_files = character(), travis = TRUE)

Arguments

name, list, environment, path, force, code_files
Same as package.skeleton.

stan_files
A character vector with paths to .stan files to include in the package (these files will be included in the src/stan_files directory). Otherwise similar to the code_files argument.

travis
Should a .travis.yml file be added to the package directory? Defaults to TRUE. The file has some settings already set to help with compilation issues, but we do not guarantee that it will work on travis-ci.

Details

This function first calls package.skeleton and then adds the files listed in stan_files to the src/stan_files directory. Finally, it downloads several files from rstanarm package's GitHub repository to facilitate building the resulting package. Note that rstanarm is licensed under the GPL >= 3, so package builders who do not want to be governed by that license should not use the downloaded files that contain R code (that said, Rcpp is GPL, so not using the rstanarm files is not the only thing impeding use of other licenses). Otherwise, it may be worth considering whether it would be easier to include your .stan programs and supporting R code in the rstanarm package.

Unlike package.skeleton, rstan_package_skeleton also creates a file in the R/ directory called "name-package.R", where name is the package name. In this file rstan_package_skeleton writes lines (using roxygen2 tags) for ensuring that some necessary content makes it into the NAMESPACE file. Before terminating, rstan_package_skeleton will run roxygen2::roxygenise so that the NAMESPACE is created.

rstan_package_skeleton will also create an RStudio project file for the package with a .Rproj extension. If not using RStudio this file can be deleted or ignored.

See also

The rstanarm repository on GitHub.

Guidelines and recommendations for developers of R packages interfacing with Stan and a demonstration getting a simple package working can be found in the vignettes included with rstantools and at http://mc-stan.org/rstantools/articles/.

After reading the guidelines for developers, if you have trouble setting up your package let us know on the the Stan Forums or at rstantools issue tracker.

The useR2016 presentation How to Use (R)Stan to Estimate Models in External R Packages.