rstantools-package {rstantools}	R Documentation

Tools for Developing R Packages Interfacing with Stan

Description

Stan Development Team

The rstantools package provides various tools for developers of R packages interfacing with Stan (https://mc-stan.org), including functions to set up the required package structure, S3 generic methods to unify function naming across Stan-based R packages, and vignettes with guidelines for developers. To get started building a package see rstan_create_package().

Author(s)

Maintainer: Jonah Gabry jgabry@gmail.com

Authors:

Ben Goodrich benjamin.goodrich@columbia.edu

Martin Lysy mlysy@uwaterloo.ca

Andrew Johnson

Other contributors:

Hamada S. Badr [contributor]

Marco Colombo [contributor]

Stefan Siegert [contributor]

Trustees of Columbia University [copyright holder]

See Also

Guidelines and recommendations for developers of R packages interfacing with Stan and a demonstration getting a simple package working can be found in the vignettes included with rstantools and at mc-stan.org/rstantools/articles.

After reading the guidelines for developers, if you have trouble setting up your package let us know on the the Stan Forums or at the rstantools GitHub issue tracker.

Useful links:

https://mc-stan.org/rstantools/

https://discourse.mc-stan.org/

Report bugs at https://github.com/stan-dev/rstantools/issues

[Package rstantools version 2.6.0 Index]