Generic function for posterior uncertainty intervals

These intervals are often referred to as credible intervals, but we use the term uncertainty intervals to highlight the fact that wider intervals correspond to greater uncertainty. See posterior_interval.stanreg in the rstanarm package for an example.

posterior_interval(object, ...)

# S3 method for default
posterior_interval(object, prob = 0.9, ...)

Arguments

object
The object to use.

...
Arguments passed to methods. See the methods in the rstanarm package for examples.

prob
A number \(p \in (0,1)\) indicating the desired probability mass to include in the intervals.

Value

posterior_interval methods should return a matrix with two columns and as many rows as model parameters (or a subset of parameters specified by the user). For a given value of prob, \(p\), the columns correspond to the lower and upper \(100p\)% interval limits and have the names \(100\alpha/2\)% and \(100(1 - \alpha/2)\)%, where \(\alpha = 1-p\). For example, if prob=0.9 is specified (a \(90\)% interval), then the column names would be "5%" and "95%", respectively.

The default method just takes object to be a matrix and computes quantiles, with prob defaulting to 0.9.

See also

The rstanarm package for example methods (CRAN, GitHub).

The guidelines for developers of R packages interfacing with Stan, a copy of which can be found in the package vignettes. See browseVignettes("rstantools") or vignette(package = "rstantools"). The document is also available online at the rstantools page on the CRAN website.

Examples

# Default method takes a numeric matrix (of posterior draws)
draws <- matrix(rnorm(100 * 5), 100, 5) # fake draws
colnames(draws) <- paste0("theta_", 1:5)
posterior_interval(draws)

#>                5%      95%
#> theta_1 -1.865365 1.683664
#> theta_2 -1.517411 1.882336
#> theta_3 -1.568718 1.618074
#> theta_4 -1.397015 1.842534
#> theta_5 -1.774352 1.521312

# Example using rstanarm package:
# posterior_interval method for 'stanreg' objects

if (require("rstanarm")) {
  fit <- stan_glmer(
    mpg ~ wt + am + (1|cyl),
    data = mtcars,
    QR = TRUE,
    prior = normal(0, 1),
    iter = 500 # to speed up example
  )
  posterior_interval(fit, prob = 0.5)
}

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Chain 1, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.115518 seconds (Warm-up)
#>                0.094151 seconds (Sampling)
#>                0.209669 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Chain 2, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 2, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 2, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 2, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 2, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 2, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 2, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 2, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 2, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 2, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 2, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 2, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.180486 seconds (Warm-up)
#>                0.121596 seconds (Sampling)
#>                0.302082 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Chain 3, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 3, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 3, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 3, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 3, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 3, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 3, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 3, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 3, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 3, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 3, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 3, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.119155 seconds (Warm-up)
#>                0.109761 seconds (Sampling)
#>                0.228916 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Chain 4, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 4, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 4, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 4, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 4, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 4, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 4, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 4, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 4, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 4, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 4, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 4, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.148191 seconds (Warm-up)
#>                0.100452 seconds (Sampling)
#>                0.248643 seconds (Total)
#> 

#> Error in posterior_interval.default(fit, prob = 0.5): For the default method 'object' should be a matrix.

# Also see help("posterior_interval", package = "rstanarm")