Generic function for accessing the posterior distribution of the linear predictor

Extract the posterior draws of the linear predictor, possibly transformed by the inverse-link function. See posterior_linpred.stanreg in the rstanarm package for an example.

posterior_linpred(object, transform = FALSE, ...)

Arguments

object
The object to use.

transform
Should the linear predictor be transformed using the inverse-link function? The default is FALSE, in which case the untransformed linear predictor is returned.

...
Arguments passed to methods. See the methods in the rstanarm package for examples.

Value

posterior_linpred methods should return a \(D\) by \(N\) matrix, where \(D\) is the number of draws from the posterior distribution distribution and \(N\) is the number of data points.

See also

The rstanarm package for example methods (CRAN, GitHub).

The guidelines for developers of R packages interfacing with Stan, a copy of which can be found in the package vignettes. See browseVignettes("rstantools") or vignette(package = "rstantools"). The document is also available online at the rstantools page on the CRAN website.

Examples

# See help("posterior_linpred", package = "rstanarm")