Generic function for predictive errors

Generic function for computing predictive errors \(y - y^{rep}\) (in-sample, for observed \(y\)) or \(y - \tilde{y}\) (out-of-sample, for new or held-out \(y\)). See predictive_error.stanreg in the rstanarm package for an example.

predictive_error(object, ...)

# S3 method for default
predictive_error(object, y, ...)

Arguments

object
The object to use.

...
Arguments passed to methods. See the methods in the rstanarm package for examples.

y
For the default method, a vector of \(y\) values the same length as the number of columns in the matrix used as object.

Value

predictive_error methods should return a \(D\) by \(N\) matrix, where \(D\) is the number of draws from the posterior predictive distribution and \(N\) is the number of data points being predicted per draw.

The default method just takes object to be a matrix and y to be a vector.

See also

The rstanarm package for example methods (CRAN, GitHub).

The guidelines for developers of R packages interfacing with Stan, a copy of which can be found in the package vignettes. See browseVignettes("rstantools") or vignette(package = "rstantools"). The document is also available online at the rstantools page on the CRAN website.

Examples

# default method
y <- rnorm(10)
ypred <- matrix(rnorm(500), 50, 10)
predictive_error(ypred, y)

#>              [,1]       [,2]        [,3]        [,4]       [,5]         [,6]
#>  [1,]  1.12338166 -0.8281493 -0.36792114  0.64814222  2.8713544 -0.112797784
#>  [2,] -0.99019108 -0.7414359 -1.97339438  1.20661238  1.2136670  1.270037359
#>  [3,] -0.28194297  0.1447368  0.97256355  1.26707042  0.4190036  0.162936853
#>  [4,] -0.06836262 -0.9223352 -0.70706187  1.17161506  1.9135477  1.120305407
#>  [5,] -0.02102343 -1.4716103 -0.16079376  2.11818963  2.6725003  0.260495027
#>  [6,]  1.20022892 -0.9298907 -2.34471983  0.38808030  3.9825416 -0.422517873
#>  [7,]  0.88835517 -0.7859034 -0.29665265  1.33217290  1.6608173 -0.366361608
#>  [8,] -0.13910686  0.4688473 -2.00479346  0.65846481  3.3162472 -0.647119614
#>  [9,] -1.03527336 -1.1829623 -0.74475024  0.98915023  2.1512549 -0.671678802
#> [10,]  0.66173637  0.4064740 -0.55601790  2.63078891  3.3794484  0.361649145
#> [11,]  1.81852138  1.2597074 -0.37741587  0.15700386 -0.0573501  0.399731768
#> [12,]  0.91369038 -0.7527352 -0.75034942 -0.32415234  0.6383193 -0.205440248
#> [13,] -0.37240357 -0.2979763  0.12405039  1.71285742  1.8684591  0.035739168
#> [14,] -1.00378468 -0.5016124 -1.18543972  1.37549351  2.1687213  0.739174982
#> [15,]  0.41570257  1.2589806 -1.09216824  2.02955679  2.4409089 -0.331697736
#> [16,]  1.42318721 -0.6034491  0.31240514  0.24718290  4.3638001  0.359242571
#> [17,]  2.51628982 -1.1005754 -0.15676320  1.36637643  3.5311401 -0.182606115
#> [18,]  0.24952622 -1.8770205 -0.55909123  0.13210803  3.1719668 -1.113559399
#> [19,]  0.92076885 -2.7930481  0.06420400  0.05742163  3.9842617 -0.435307997
#> [20,]  1.30535250 -2.0641912 -0.45972473  1.49856875  1.6686682 -0.004100808
#> [21,]  1.14988456 -2.4365829  0.77899711  0.41154171  3.7828095 -2.130241719
#> [22,]  1.18647368 -3.5414512 -1.47014443  2.16690918  1.3852179 -0.560590255
#> [23,]  0.95957953 -2.0572557 -1.53161104  0.11322571  2.8752941 -1.038761582
#> [24,]  1.79074896  0.1943873  1.66386307  0.23449348  1.8570930 -0.637366368
#> [25,]  0.68252826 -2.0599448 -0.19921396  0.35519549  1.2484746  1.320184237
#> [26,]  0.88018265 -1.0383827 -0.68183022  0.49809677  3.0898687  0.671120087
#> [27,]  1.36751867 -0.7501838 -1.92995800  0.49368227  1.4241451  0.481227000
#> [28,] -1.37804332 -1.9352874  0.11738441  3.01181718  2.0991423 -0.117605175
#> [29,]  0.46898669 -1.0575496 -0.37071973  1.10249507  2.5431770 -0.332802201
#> [30,]  0.82608590 -1.6480557 -0.77011393  0.58418724  2.0761080 -0.065303543
#> [31,]  2.65130138 -0.1683289  0.95421633  2.31670165  2.3347424 -0.251038667
#> [32,]  1.31406861 -1.2800719 -0.73690270  0.13711616  4.6113014  0.207531819
#> [33,]  1.20042224 -2.3308853  0.44896587 -0.11853348  3.3552750 -0.346374755
#> [34,]  1.15320356 -0.7045122 -2.40100159  0.58300653  2.0880544  0.741764199
#> [35,] -0.26075152 -0.9214840  0.54020836  1.11028995  2.3176243 -0.756199365
#> [36,] -1.32779798 -0.8516938  0.83118410  0.53830072  2.3332244 -0.957555445
#> [37,] -0.38278092 -1.6138894 -2.26728699 -1.08100531  1.2810440 -0.241914596
#> [38,]  1.18171606 -0.3432961 -1.00862561  2.23668528  0.7138631 -0.097205013
#> [39,]  0.52569520 -0.8805091  0.15273003  1.20667962  0.5452384 -0.598330260
#> [40,] -0.61533841  0.4041412 -1.17007102 -0.04485427  2.3714740 -0.214137615
#> [41,]  0.10522138 -1.4217419 -1.87543833  2.57002502  1.2146246 -0.132242164
#> [42,]  0.69635645 -1.3003897 -1.09714416  0.90218066  2.3780837  0.345882564
#> [43,]  1.22218748 -1.1791382 -1.31186962 -0.22364618  2.2255794 -0.521472751
#> [44,] -1.29164842 -0.5287655  0.42678953  1.01061101  0.7027488 -0.582988348
#> [45,]  1.87644336 -1.7970626  0.46632039  0.90245452  2.3813203 -0.744247473
#> [46,] -0.52979591 -1.2595703  0.06092086  1.31221395  3.2807906  1.310116442
#> [47,]  1.01372975 -0.9905371 -2.39105089  0.48122488  1.8203915 -1.488847443
#> [48,] -0.21770044 -2.9947141 -0.38579258  0.57187639  3.2192511 -1.195663678
#> [49,]  1.23051283 -1.2983365 -0.45278374  2.16456234  2.1765652 -0.482416105
#> [50,]  1.53327358 -0.1383649 -0.91509308  0.95370311  2.7117796  0.946861576
#>               [,7]        [,8]        [,9]       [,10]
#>  [1,] -1.307251969  0.18318574  1.60840099 -0.76454026
#>  [2,] -2.495149352  0.65047069 -0.71033688 -0.81997941
#>  [3,] -1.800026806  0.91853982 -0.18298866 -0.56671980
#>  [4,] -1.151391334  1.14954252  0.58465937 -1.28710415
#>  [5,]  0.382588929  1.20501103  0.63027819 -1.01727857
#>  [6,] -0.941981084  2.92553521 -0.46482724 -0.53144342
#>  [7,]  0.091178134  1.29901963 -0.21870870  0.60918353
#>  [8,]  0.264277437  1.39569676 -0.52827195 -1.48023907
#>  [9,] -1.916557132  0.65524171  1.00114386 -1.49866010
#> [10,] -0.395719470  1.23057880 -0.42459867 -1.35238357
#> [11,]  0.559208785  0.42949484  2.25307343 -0.25706497
#> [12,]  0.239069823  2.04573396 -1.31312786 -1.16565702
#> [13,] -0.334282529  1.16387420 -1.37062209 -2.52984002
#> [14,] -0.644409972  2.39355090  0.26793929 -1.10847422
#> [15,] -1.415338222  0.54210064 -0.27089958  0.69073169
#> [16,] -0.979443829 -0.48717178 -0.65023856 -0.87243374
#> [17,] -4.272986276  1.42354040  1.38941548 -0.65042309
#> [18,] -1.441783806 -0.05098877  0.91909954  0.02041591
#> [19,] -1.086168103  2.10510006  0.35196919 -0.69133705
#> [20,] -1.515386513 -0.67938971  0.04372814 -0.86885464
#> [21,] -0.380904557  1.04884756 -1.07132425 -1.52347103
#> [22,] -1.396144122  1.71460618 -0.97546797 -1.71697471
#> [23,] -1.124654617  1.67767821  0.82715041 -2.30585904
#> [24,] -1.044264497  0.12569273  0.29026411  1.32370283
#> [25,] -0.561803149 -1.26163703  0.15168569 -0.42880448
#> [26,]  0.007175285  1.61244604  0.84229024 -0.52862858
#> [27,] -1.260387328  1.36298901  0.85255182 -1.89539796
#> [28,] -0.468881285  1.71914853 -2.19400762 -1.75590275
#> [29,] -3.723621963  0.37985704  0.52359559 -0.81271270
#> [30,]  0.905626331  3.16390647  1.42386401 -0.63625391
#> [31,] -0.012351564  1.63681308 -1.18098094 -0.21268518
#> [32,] -1.414613605  1.32494937  0.55905934 -0.26143246
#> [33,]  0.465521435 -1.11716757  1.28944220 -1.80394609
#> [34,]  0.464734413  2.04938767  1.97677264 -1.34369317
#> [35,] -0.553965522  1.89733818 -0.29889727 -1.24925773
#> [36,] -0.104994442  1.57716783 -0.82604062  0.22505967
#> [37,] -2.347390832  0.17464196 -0.65865601 -1.13877089
#> [38,] -1.853989778  0.25444131  1.36260081  0.83275551
#> [39,] -1.455926262 -0.78002728 -0.05382684 -1.16550582
#> [40,] -0.522924992  0.09009016 -2.63278412  1.60188123
#> [41,] -1.961306146  0.51149113  0.39950059 -1.63935206
#> [42,] -0.866928784  1.99667715  1.28181058 -2.06685007
#> [43,] -2.160401322  0.98728038  0.31251293 -0.63296122
#> [44,]  0.831393924  1.07328079 -1.17864331 -1.14137054
#> [45,] -0.990441156  1.26086297 -0.04602281 -2.80960044
#> [46,] -0.793508187  0.46779325  0.19027900 -1.18064557
#> [47,]  0.538751773  0.12888331 -0.45129869 -1.05892463
#> [48,] -1.134560779  0.68478708 -0.18205546 -2.01049660
#> [49,] -1.194111751  3.09016525  0.73438113 -0.82229974
#> [50,] -1.036253441  0.25129086 -1.60877934 -0.98578814

# Also see help("predictive_error", package = "rstanarm")