rstantools-package {rstantools}	R Documentation

Tools for Developing R Packages Interfacing with Stan

Description

Stan Development Team

The rstantools package provides various tools for developers of R packages interfacing with Stan (http://mc-stan.org), including functions to set up the required package structure, S3 generic methods to unify function naming across Stan-based R packages, and a vignette with guidelines for developers. To get started building a package see rstan_package_skeleton.

See Also

The guidelines for developers of R packages interfacing with Stan, a copy of which can be found in the package vignettes. See browseVignettes("rstantools") or vignette(package = "rstantools"). The document is also available online at the rstantools page on the CRAN website.

The Stan Forums and the rstantools issue tracker for assistance if you have trouble setting up your package.

[Package rstantools version 1.4.0 Index]